roll back emsdk to build correct samples

+++ upython.c
@@ -39,6 +39,8 @@ Py_InitializeEx(int param) {
     #endif
 
     #if MICROPY_ENABLE_PYSTACK
+    // thx @embeddedt
+    // https://github.com/littlevgl/lv_micropython/commit/19b2bec8863924c8a46f7761c4568dea2873670d
         mp_pystack_init(pystack, &pystack[MP_ARRAY_SIZE(pystack)]);
     #endif
 
@@ -134,14 +136,14 @@ mp_lexer_new_from_file(const char *filename) {
     }
     fseeko(file, 0, SEEK_END);
     off_t size_of_file = ftello(file);
-    fprintf(stderr, "mp_lexer_new_from_file(%s size=%lld)\n", filename, size_of_file );
+    fprintf(stderr, "mp_lexer_new_from_file(%s size=%lld)\n", filename, (long long)size_of_file );
     fseeko(file, 0, SEEK_SET);
 
     char * code_buf = malloc(size_of_file);
     fread(code_buf, size_of_file, 1, file);
 
     if (code_buf == NULL) {
-        fprintf(stderr, "READ ERROR: mp_lexer_new_from_file(%s size=%lld)\n", filename, size_of_file );
+        fprintf(stderr, "READ ERROR: mp_lexer_new_from_file(%s size=%lld)\n", filename, (long long)size_of_file );
         return NULL;
     }
     //fprintf(stderr, "%s" ,  code_buf);
